feat: Address multiple UI and build issues
This commit addresses three separate issues reported by the user:

1.  **fix(middleware): Exclude static files from auth middleware**
    - Updates the middleware matcher in `src/middleware.ts` to exclude paths containing a dot (e.g., .png, .svg).
    - This prevents the authentication middleware from intercepting requests for static assets, allowing them to be served correctly on unauthenticated pages.

2.  **style(dialog): Make project dialog responsive**
    - Modifies the Tailwind CSS classes for the `DialogContent` in `src/components/projects/manage-project-dialog.tsx`.
    - The dialog now has a responsive width that adjusts to different screen sizes and its content is scrollable, preventing layout breakage.

3.  **fix(build): Resolve useSearchParams error on login page**
    - Refactors the login page (`src/app/login/page.tsx`) to use a Suspense boundary.
    - The page logic was extracted into a new client component (`src/app/login/login-page-client.tsx`), which is then rendered within `<Suspense>` in the main page component.
    - This resolves the build error `useSearchParams() should be wrapped in a suspense boundary`.
