+++ src/app/login/page.tsx
@@ -1,115 +1,48 @@
-'use client';
-import { useState, useEffect } from 'react';
-import { Button } from '@/components/ui/button';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { useRouter, useSearchParams } from 'next/navigation';
-import { useToast } from '@/hooks/use-toast';
-import { signIn } from 'next-auth/react';
-import { Separator } from '@/components/ui/separator';
+import { Suspense } from 'react';
+import LoginPageClient from './login-page-client';
+import { Skeleton } from '@/components/ui/skeleton';
 
-export default function LoginPage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const { toast } = useToast();
-  const [email, setEmail] = useState('admin@chb.com.br');
-  const [password, setPassword] = useState('chb123');
-  const [isLoading, setIsLoading] = useState(false);
-  const [isGoogleLoading, setIsGoogleLoading] = useState(false);
-
-  useEffect(() => {
-    const error = searchParams.get('error');
-    if (error === 'AccessDenied') {
-      toast({
-        variant: 'destructive',
-        title: 'Acesso Negado',
-        description: 'Seu email do Google não corresponde a nenhum usuário existente no sistema.',
-      });
-    }
-  }, [searchParams, toast]);
-
-  const handleLogin = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-
-    const result = await signIn('credentials', {
-      redirect: false,
-      email,
-      password,
-    });
-
-    if (result?.ok) {
-      router.push('/');
-    } else {
-      toast({
-        variant: 'destructive',
-        title: 'Falha no login',
-        description: result?.error || 'Credenciais inválidas. Verifique seu email e senha.',
-      });
-      setIsLoading(false);
-    }
-  };
-
-  const handleGoogleLogin = async () => {
-    setIsGoogleLoading(true);
-    await signIn('google', { callbackUrl: '/' });
-    // No need to set loading to false, as the page will redirect.
-  };
-
+// A simple skeleton loader for the login page
+function LoginPageSkeleton() {
   return (
     <div className="flex items-center justify-center min-h-screen bg-background">
-      <Card className="w-full max-w-sm">
-        <CardHeader>
-          <CardTitle className="text-2xl font-headline">Login</CardTitle>
-          <CardDescription>
-            Entre com seu email e senha para acessar o painel.
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <form onSubmit={handleLogin} className="space-y-4">
+      <div className="w-full max-w-sm p-8 space-y-6">
+        <div className="space-y-2 text-center">
+          <Skeleton className="h-8 w-3/4 mx-auto" />
+          <Skeleton className="h-4 w-full mx-auto" />
+        </div>
+        <div className="space-y-4">
             <div className="space-y-2">
-              <Label htmlFor="email">Email</Label>
-              <Input
-                id="email"
-                type="email"
-                placeholder="m@example.com"
-                required
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                disabled={isLoading}
-              />
+                <Skeleton className="h-4 w-1/4" />
+                <Skeleton className="h-10 w-full" />
             </div>
             <div className="space-y-2">
-              <Label htmlFor="password">Senha</Label>
-              <Input
-                id="password"
-                type="password"
-                required
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                disabled={isLoading}
-              />
+                <Skeleton className="h-4 w-1/4" />
+                <Skeleton className="h-10 w-full" />
             </div>
-            <Button type="submit" className="w-full" disabled={isLoading}>
-              {isLoading ? 'Entrando...' : 'Entrar com Email'}
-            </Button>
-          </form>
-          <div className="relative my-4">
+            <Skeleton className="h-10 w-full" />
+        </div>
+        <div className="relative my-4">
             <div className="absolute inset-0 flex items-center">
-              <Separator />
+                <Skeleton className="h-px w-full" />
             </div>
             <div className="relative flex justify-center text-xs uppercase">
-              <span className="bg-background px-2 text-muted-foreground">
-                Ou continue com
-              </span>
+                <span className="bg-background px-2 text-muted-foreground">
+                    Ou continue com
+                </span>
             </div>
-          </div>
-          <Button variant="outline" className="w-full" onClick={handleGoogleLogin} disabled={isGoogleLoading || isLoading}>
-            {isGoogleLoading ? 'Redirecionando...' : 'Login com Google'}
-          </Button>
-        </CardContent>
-      </Card>
+        </div>
+        <Skeleton className="h-10 w-full" />
+      </div>
     </div>
+  );
+}
+
+
+export default function LoginPage() {
+  return (
+    <Suspense fallback={<LoginPageSkeleton />}>
+      <LoginPageClient />
+    </Suspense>
   );
 }
